Change daemon user from nobody to kurento
Create (if not exist) a new user kurento and
change the daemon user from nobody to kurento. This change is
necessary to avoid errors in playerendpoint. The configuration of
the internal uridecodebin in the playerendpoint was changed to
support http seeks (property download" was setted to TRUE) and
now it needs to save temporal buffers and need write permissions,
for that reason we can not use longer the nobody user.

Change-Id: Ib3a3eb4ca7d4149c4f219017e3865a74448f0681